cli: fix skipping columns with all default-values
The header-cell is the header for a column with possibly many
rows.

We must not set:
  header_cell->skip = nmc_config->overview && is_default;
for a particular cell, because it does not take into accound
the neighbouring rows. It is only correct, if there is only
one row.

Reverse the logic and rename "strip" to "to_print".
The cell then can opt-in, whether it needs to be printed.
If not cell opts-in, it will be skipped.

Fixes: 9a19bbcb2f6b78be2541334a0d650e765a11fa2b